fix warning: use size_t when comparing to SK_ARRAY_COUNT
git-svn-id: http://skia.googlecode.com/svn/trunk@8108 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/core/SkPictureRecord.cpp
@@ -529,7 +529,7 @@ void SkPictureRecord::restore() {
     }
 
     uint32_t initialOffset, size;
-    int opt;
+    size_t opt;
     for (opt = 0; opt < SK_ARRAY_COUNT(gPictureRecordOpts); ++opt) {
         if ((*gPictureRecordOpts[opt])(&fWriter, fRestoreOffsetStack.top(), &fPaints)) {
             // Some optimization fired so don't add the RESTORE
